+++ test/vine_test.dart
@@ -7,6 +7,7 @@ import './rules/string_test.dart' as string_test;
 import './rules/object_test.dart' as object_test;
 import './rules/union_test.dart' as union_test;
 import './rules/date_test.dart' as date_test;
+import './open_api_test.dart' as openapi_test;
 
 void main() async {
   string_test.main();
@@ -18,4 +19,5 @@ void main() async {
   object_test.main();
   union_test.main();
   date_test.main();
+  openapi_test.main();
 }